Ai/image_toxicity/requirements.txt
torch
torchvision
transformers
pillow
onnx
onnxruntime
numpy
